20220224/requirements.txt
flask
flask_restful
google-cloud-firestore
google-cloud-pubsub
gunicorn
requests
